test(fulgur): move E2E render smoke tests to tests/render_smoke.rs
Extract 10 end-to-end smoke tests from engine.rs::tests into a dedicated
integration test file. engine.rs::tests now holds only Engine builder /
template unit tests. Update CLAUDE.md Gotchas to point at the new home so
future draw paths land in the right file by default.

Closes fulgur-181g
